requirements.txt
boto3~=1.20.24

botocore~=1.23.45

celery>=5.2.0, <5.2.3

coverage>=5.5,<6.0

ddt~=1.4.2 , <1.5.0

Django>=3.2.3,<4.0.0

django-celery-beat >=2.2.1, <2.3.0

django-celery-results >=2.2.0, <2.3.0

django-model-utils>=4.1.1,<4.2.0

django-mysql>=4.4.0,<4.5.0

djangorestframework>=3.13.0,<3.13.1

djangorestframework-xml==2.0.0

django-ses>=2.2.0 , <2.3.1

flake8>=3.9.0,<4.0.1

flower <=1.0.0

fsspec>=2022.1.0

gunicorn>=20.1.0,<20.2.0

martor>=1.6.0,<1.6.10

mysql-connector-python >= 8.0.28, <=8.0.29

numpy>=1.21.1

openlxp-notification==1.3.7

openlxp-xia==1.3.2

openpyxl >=3.0.7 , <3.1.0

pandas>=1.3.5,<1.4.0

Pillow >=9.0.0

pytz >=2021.1, <2021.3

redis==4.1.2

requests~=2.27.1

s3fs==0.4.2

xlrd >=2.0.1 , <2.1.0
